pyproject.toml
[tool.poetry]
name = "devops-course-starter"
version = "0.1.0"
description = ""
authors = ["Joel Wee <joel.wee@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
gunicorn = "^20.0.4"
pyyaml = "^5.3.1"
pymongo = "^3.11.3"
dnspython = "^2.1.0"
mockupdb = "^1.8.0"
flask-pymongo = "^2.3.0"

[tool.poetry.dev-dependencies]
selenium = "^3.141.0"
rope = "^0.17.0"
pytest = "^6.0.1"
flake8 = "^3.8.3"
watchdog = "^0.10.3"
autopep8 = "^1.5.5"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
